backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
twilio
apscheduler
requests
SQLAlchemy
pydantic
PyJWT
bcrypt 